Shifting components into a dedicated hud

+++ lib/helpers/timestep/debug/timestep_faker.dart
@@ -8,7 +8,7 @@ class TimestepFaker extends Component {
   static const componentKey = 'TimestepFaker';
 
   final bool useFakeTimestep;
-  final double fakeFps;
+  final int fakeFps;
   late final double fakeTimestep;
 
   double _fpsFrameCounter = 0;
